Revert "Rename triton function"
