+++ flash_sparse_attn/ops/triton/flash_dense_fwd.py
@@ -7,7 +7,6 @@ import triton.language as tl
 
 from flash_sparse_attn.ops.triton import (
     assert_inputs,
-    flash_decode_combine,
     utils,
     launch_template,
     launch_grid,
@@ -15,6 +14,7 @@ from flash_sparse_attn.ops.triton import (
     block_info,
     activations,
     mask,
+    flash_fwd_combine,
 )
 
 
@@ -743,7 +743,7 @@ def _flash_dense_attn_base_forward(
     )
 
     if is_split_kv:
-        flash_decode_combine._flash_attn_fwd_combine(
+        flash_fwd_combine._flash_attn_fwd_combine(
             out_partial,
             lse_partial,
             out,
@@ -894,7 +894,7 @@ def _flash_dense_attn_varlen_base_forward(
     )
 
     if is_split_kv:
-        flash_decode_combine._flash_attn_fwd_combine(
+        flash_fwd_combine._flash_attn_fwd_combine(
             out_partial,
             lse_partial,
             out,
