+++ flash_sparse_attn/ops/triton/flash_sparse_fwd.py
@@ -7,7 +7,6 @@ import triton.language as tl
 
 from flash_sparse_attn.ops.triton import (
     assert_inputs,
-    flash_decode_combine,
     utils,
     launch_template,
     launch_grid,
@@ -15,27 +14,21 @@ from flash_sparse_attn.ops.triton import (
     block_info,
     activations,
     mask,
+    flash_fwd_combine,
 )
 
 
 @triton.jit
-def _fwd_inner_gated_base_kernel(
-    skip_gate_curr,
-    acc_s,
-    acc_o,
+def _fwd_inner_sparse_base_kernel(
     q_tile,
-    a_tile,
+    k_tile,
     k_ptrs,
     v_ptrs,
-    d_ptrs,
-    a_max,
-    a_min,
-    gate_max,
+    acc_o,
     block_max,
     row_max,
     row_sum,
     softmax_scale_log2,
-    gate_threshold_log2,
     softmax_threshold_log2,
     m_block,
     n_block,
@@ -50,135 +43,75 @@ def _fwd_inner_gated_base_kernel(
     IS_MASK: tl.constexpr,
     MASK_CAUSAL: tl.constexpr,
     MASK_LOCAL: tl.constexpr,
-    IS_LOGSIGMOID_GATE: tl.constexpr,
     CHECK_INF: tl.constexpr,
 ):
-    # Advance delta pointers
-    d_ptrs = tl.advance(d_ptrs, (-TILE_N,))
+    # Compute attention scores
+    acc_s = tl.dot(q_tile, k_tile)
 
-    # Load next delta tile
-    d_tile = tl.load(d_ptrs, boundary_check=(0,)).to(tl.float32)
-    d_max = tl.max(d_tile)
-    d_min = tl.min(d_tile)
-
-    skip_gate_next = True
-    if not skip_gate_curr:
-        # Advance key pointers
-        k_ptrs = tl.advance(k_ptrs, (0, -TILE_N))
-
-        if n_block > n_block_min:
-            # Check if any gates are active for next tile
-            gate_max, skip_gate_next = activations.online_gate(
-                a_max,
-                a_min,
-                d_max,
-                d_min,
-                gate_max,
-                scale_log2=softmax_scale_log2,
-                gate_threshold_log2=gate_threshold_log2,
-            )
-
-        if IS_MASK:
-            # Apply mask
-            acc_s = mask.apply_mask(
-                acc_s=acc_s,
-                m_block=m_block,
-                n_block=n_block,
-                seqlen_q=actual_seqlen_q,
-                seqlen_k=actual_seqlen_k,
-                MASK_SEQLEN=True,
-                MASK_CAUSAL=MASK_CAUSAL,
-                MASK_LOCAL=MASK_LOCAL,
-                TILE_M=TILE_M,
-                TILE_N=TILE_N,
-                WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
-                WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
-                QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-                SWAP_AB=False,
-            )
-
-        # Apply online softmax
-        p, block_max, row_max, row_sum, row_scale, skip_softmax = (
-            activations.online_softmax(
-                acc_s=acc_s,
-                block_max=block_max,
-                row_max=row_max,
-                row_sum=row_sum,
-                scale_log2=softmax_scale_log2,
-                softmax_threshold_log2=softmax_threshold_log2,
-                CHECK_INF=CHECK_INF,
-                RESCALE_THRESHOLD=0.0,
-            )
-        )
-
-        if not skip_softmax:
-            # Load value tile
-            v_tile = tl.load(v_ptrs, boundary_check=(0, 1))
-
-            # Rescale output accumulator
-            acc_o = activations.rescale_o(acc_o, row_scale, LAZY_RESCALE=False)
-
-            # Update output accumulator
-            acc_o += tl.dot(p.to(v_tile.dtype), v_tile)
-
-        # Advance value pointer
-        v_ptrs = tl.advance(v_ptrs, (-TILE_N, 0))
-    else:
-        # Advance key and value pointers
-        k_ptrs = tl.advance(k_ptrs, (0, -TILE_N))
-        v_ptrs = tl.advance(v_ptrs, (-TILE_N, 0))
-
-        if n_block > n_block_min:
-            # Check if any gates are active for next tile
-            gate_max, skip_gate_next = activations.online_gate(
-                a_max,
-                a_min,
-                d_max,
-                d_min,
-                gate_max,
-                scale_log2=softmax_scale_log2,
-                gate_threshold_log2=gate_threshold_log2,
-            )
-
-    if not skip_gate_next:
-        # Compute attention gates for next tile
-        acc_s = a_tile[:, None] * d_tile[None, :]
-
-        if IS_LOGSIGMOID_GATE:
-            acc_s = activations.log_sigmoid(acc_s, FASTMATH=True)
-
+    # Advance key pointer
+    k_ptrs = tl.advance(k_ptrs, (0, -TILE_N))
+    if n_block > n_block_min:
         # Load next key tile
         k_tile = tl.load(k_ptrs, boundary_check=(0, 1))
 
-        # Compute attention scores for next tile
-        acc_s += tl.dot(q_tile, k_tile)
+    if IS_MASK:
+        # Apply mask to attention scores
+        acc_s = mask.apply_mask(
+            acc_s=acc_s,
+            m_block=m_block,
+            n_block=n_block,
+            seqlen_q=actual_seqlen_q,
+            seqlen_k=actual_seqlen_k,
+            MASK_SEQLEN=True,
+            MASK_CAUSAL=MASK_CAUSAL,
+            MASK_LOCAL=MASK_LOCAL,
+            TILE_M=TILE_M,
+            TILE_N=TILE_N,
+            WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
+            WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
+            QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
+            SWAP_AB=False,
+        )
 
-    return (
-        skip_gate_next,
-        acc_s,
-        acc_o,
-        k_ptrs,
-        v_ptrs,
-        d_ptrs,
-        gate_max,
-        block_max,
-        row_max,
-        row_sum,
+    # Apply online softmax
+    p, block_max, row_max, row_sum, row_scale, skip_softmax = (
+        activations.online_softmax(
+            acc_s=acc_s,
+            block_max=block_max,
+            row_max=row_max,
+            row_sum=row_sum,
+            scale_log2=softmax_scale_log2,
+            softmax_threshold_log2=softmax_threshold_log2,
+            CHECK_INF=CHECK_INF,
+            RESCALE_THRESHOLD=0.0,
+        )
     )
+
+    if not skip_softmax:
+        # Load value tile
+        v_tile = tl.load(v_ptrs, boundary_check=(0, 1))
+
+        # Rescale output accumulator
+        acc_o = activations.rescale_o(acc_o, row_scale, LAZY_RESCALE=False)
+
+        # Update output accumulator
+        acc_o += tl.dot(p.to(v_tile.dtype), v_tile)
+
+    # Advance value pointer
+    v_ptrs = tl.advance(v_ptrs, (-TILE_N, 0))
+
+    return k_tile, k_ptrs, v_ptrs, acc_o, block_max, row_max, row_sum
 
 
 @triton.jit
-def _fwd_gated_base_kernel(
+def _fwd_base_sparse_kernel(
     Q,
     K,
     V,
-    A,
-    D,
     Out,
     Lse,
     softmax_scale_log2,
     softmax_threshold,
-    gate_threshold,
     stride_qb,
     stride_qh,
     stride_qm,
@@ -188,10 +121,6 @@ def _fwd_gated_base_kernel(
     stride_vb,
     stride_vh,
     stride_vn,
-    stride_ab,
-    stride_ah,
-    stride_db,
-    stride_dh,
     stride_ob,
     stride_oh,
     stride_om,
@@ -222,8 +151,6 @@ def _fwd_gated_base_kernel(
     HAS_SEQUSED_Q: tl.constexpr,
     HAS_SEQUSED_K: tl.constexpr,
     PACK_GQA: tl.constexpr,
-    IS_LOGSIGMOID_GATE: tl.constexpr,
-    IS_ADAPT_GATE: tl.constexpr,
 ):
     m_block = tl.program_id(0)
     head_idx = tl.program_id(1)
@@ -294,26 +221,6 @@ def _fwd_gated_base_kernel(
         padded_offset_k,
         stride_vb,
         stride_vn,
-        HAS_CU_SEQLENS_K,
-        USE_PADDED=False,
-    )
-    a_base = seqlen_info.offset_batch_Q(
-        A + head_idx * stride_ah if not PACK_GQA else A,
-        batch_idx,
-        offset_q,
-        padded_offset_q,
-        stride_ab,
-        1,
-        HAS_CU_SEQLENS_Q,
-        USE_PADDED=False,
-    )
-    d_base = seqlen_info.offset_batch_K(
-        D + head_kv_idx * stride_dh,
-        batch_idx,
-        offset_k,
-        padded_offset_k,
-        stride_db,
-        1,
         HAS_CU_SEQLENS_K,
         USE_PADDED=False,
     )
@@ -462,14 +369,6 @@ def _fwd_gated_base_kernel(
             block_shape=(TILE_M, TILE_K),
             order=(1, 0),
         )
-        a_ptrs = tl.make_block_ptr(
-            base=a_base,
-            shape=(actual_seqlen_q,),
-            strides=(1,),
-            offsets=(m_block * TILE_M,),
-            block_shape=(TILE_M,),
-            order=(0,),
-        )
     else:
         q_ptrs = seqlen_info.make_pack_gqa_ptrs(
             q_base,
@@ -479,16 +378,6 @@ def _fwd_gated_base_kernel(
             stride_qm,
             TILE_M=TILE_M,
             TILE_K=TILE_K,
-            QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-        )
-        a_ptrs = seqlen_info.make_pack_gqa_ptrs(
-            a_base,
-            m_block,
-            head_idx,
-            stride_ah,
-            1,
-            TILE_M=TILE_M,
-            TILE_K=1,
             QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
         )
     k_ptrs = tl.make_block_ptr(
@@ -506,26 +395,6 @@ def _fwd_gated_base_kernel(
         offsets=((n_block_max - 1) * TILE_N, 0),
         block_shape=(TILE_N, TILE_K),
         order=(1, 0),
-    )
-    d_ptrs = tl.make_block_ptr(
-        base=d_base,
-        shape=(actual_seqlen_k,),
-        strides=(1,),
-        offsets=((n_block_max - 1) * TILE_N,),
-        block_shape=(TILE_N,),
-        order=(0,),
-    )
-
-    # Get gate threshold
-    gate_threshold_log2 = seqlen_info.get_gate_threshold(
-        gate_threshold=gate_threshold,
-        m_block=m_block,
-        seqlen_q=actual_seqlen_q,
-        seqlen_k=actual_seqlen_k,
-        IS_CAUSAL=IS_CAUSAL,
-        TILE_M=TILE_M,
-        QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-        IS_ADAPT_GATE=IS_ADAPT_GATE,
     )
 
     # Get softmax threshold
@@ -550,91 +419,65 @@ def _fwd_gated_base_kernel(
     else:
         q_tile = tl.load(q_ptrs, boundary_check=(0, 1))
 
-    # Load key tile
-    k_tile = tl.load(k_ptrs, boundary_check=(0, 1))
-
     # Initialize accumulators
-    gate_max = tl.full((), float("-inf"), dtype=tl.float32)
     block_max = tl.full((), float("-inf"), dtype=tl.float32)
     row_max = tl.full((TILE_M,), float("-inf"), dtype=tl.float32)
     row_sum = tl.zeros((TILE_M,), dtype=tl.float32)
     acc_o = tl.zeros((TILE_M, TILE_K), dtype=tl.float32)
 
-    # Load alpha tile
-    if PACK_GQA:
-        a_tile = tl.load(
-            a_ptrs,
-            mask=((offs_m // QHEADS_PER_KVHEAD_PACKGQA) < actual_seqlen_q),
-            other=0.0,
-        ).to(tl.float32)
-    else:
-        a_tile = tl.load(a_ptrs, boundary_check=(0,)).to(tl.float32)
-    a_max = tl.max(a_tile)
-    a_min = tl.min(a_tile)
-
-    # Load delta tile
-    d_tile = tl.load(d_ptrs, boundary_check=(0,)).to(tl.float32)
-    d_max = tl.max(d_tile)
-    d_min = tl.min(d_tile)
-
-    # Check if any gates are active for current tile
-    gate_max, skip_gate_curr = activations.online_gate(
-        a_max,
-        a_min,
-        d_max,
-        d_min,
-        gate_max,
-        scale_log2=softmax_scale_log2,
-        gate_threshold_log2=gate_threshold_log2,
-    )
-
-    # Initialize skip_gate_curr to False for the first iteration
-    skip_gate_curr = False
-
-    # Compute attention gates for first tile
-    acc_s = a_tile[:, None] * d_tile[None, :]
-
-    if IS_LOGSIGMOID_GATE:
-        acc_s = activations.log_sigmoid(acc_s, FASTMATH=True)
-
-    # Compute attention scores for first tile
-    acc_s += tl.dot(q_tile, k_tile)
+    # Load key tile
+    k_tile = tl.load(k_ptrs, boundary_check=(0, 1))
 
     # Process n_blocks with masking
     if IS_CAUSAL or IS_LOCAL:
         for n_block in tl.range(n_block_max - 1, n_block_max_no_mask - 1, -1):
-            (
-                skip_gate_curr,
-                acc_s,
-                acc_o,
-                k_ptrs,
-                v_ptrs,
-                d_ptrs,
-                gate_max,
-                block_max,
-                row_max,
-                row_sum,
-            ) = _fwd_inner_gated_base_kernel(
-                skip_gate_curr=skip_gate_curr,
-                acc_s=acc_s,
-                acc_o=acc_o,
+            k_tile, k_ptrs, v_ptrs, acc_o, block_max, row_max, row_sum = (
+                _fwd_inner_sparse_base_kernel(
+                    q_tile=q_tile,
+                    k_tile=k_tile,
+                    k_ptrs=k_ptrs,
+                    v_ptrs=v_ptrs,
+                    acc_o=acc_o,
+                    block_max=block_max,
+                    row_max=row_max,
+                    row_sum=row_sum,
+                    softmax_scale_log2=softmax_scale_log2,
+                    softmax_threshold_log2=softmax_threshold_log2,
+                    m_block=m_block,
+                    n_block=n_block,
+                    n_block_min=n_block_max_no_mask,
+                    actual_seqlen_q=actual_seqlen_q,
+                    actual_seqlen_k=actual_seqlen_k,
+                    TILE_M=TILE_M,
+                    TILE_N=TILE_N,
+                    WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
+                    WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
+                    QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
+                    IS_MASK=True,
+                    MASK_CAUSAL=IS_CAUSAL,
+                    MASK_LOCAL=IS_LOCAL,
+                    CHECK_INF=True,
+                )
+            )
+    else:
+        # First iteration with seqlen masking
+        n_block = n_block_max - 1
+
+        k_tile, k_ptrs, v_ptrs, acc_o, block_max, row_max, row_sum = (
+            _fwd_inner_sparse_base_kernel(
                 q_tile=q_tile,
-                a_tile=a_tile,
+                k_tile=k_tile,
                 k_ptrs=k_ptrs,
                 v_ptrs=v_ptrs,
-                d_ptrs=d_ptrs,
-                a_max=a_max,
-                a_min=a_min,
-                gate_max=gate_max,
+                acc_o=acc_o,
                 block_max=block_max,
                 row_max=row_max,
                 row_sum=row_sum,
                 softmax_scale_log2=softmax_scale_log2,
-                gate_threshold_log2=gate_threshold_log2,
                 softmax_threshold_log2=softmax_threshold_log2,
                 m_block=m_block,
                 n_block=n_block,
-                n_block_min=n_block_max_no_mask,
+                n_block_min=n_block,
                 actual_seqlen_q=actual_seqlen_q,
                 actual_seqlen_k=actual_seqlen_k,
                 TILE_M=TILE_M,
@@ -643,59 +486,10 @@ def _fwd_gated_base_kernel(
                 WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
                 QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
                 IS_MASK=True,
-                MASK_CAUSAL=IS_CAUSAL,
-                MASK_LOCAL=IS_LOCAL,
-                IS_LOGSIGMOID_GATE=IS_LOGSIGMOID_GATE,
+                MASK_CAUSAL=False,
+                MASK_LOCAL=False,
                 CHECK_INF=True,
             )
-    else:
-        # First iteration with seqlen masking
-        n_block = n_block_max - 1
-
-        (
-            skip_gate_curr,
-            acc_s,
-            acc_o,
-            k_ptrs,
-            v_ptrs,
-            d_ptrs,
-            gate_max,
-            block_max,
-            row_max,
-            row_sum,
-        ) = _fwd_inner_gated_base_kernel(
-            skip_gate_curr=skip_gate_curr,
-            acc_s=acc_s,
-            acc_o=acc_o,
-            q_tile=q_tile,
-            a_tile=a_tile,
-            k_ptrs=k_ptrs,
-            v_ptrs=v_ptrs,
-            d_ptrs=d_ptrs,
-            a_max=a_max,
-            a_min=a_min,
-            gate_max=gate_max,
-            block_max=block_max,
-            row_max=row_max,
-            row_sum=row_sum,
-            softmax_scale_log2=softmax_scale_log2,
-            gate_threshold_log2=gate_threshold_log2,
-            softmax_threshold_log2=softmax_threshold_log2,
-            m_block=m_block,
-            n_block=n_block,
-            n_block_min=n_block_min,
-            actual_seqlen_q=actual_seqlen_q,
-            actual_seqlen_k=actual_seqlen_k,
-            TILE_M=TILE_M,
-            TILE_N=TILE_N,
-            WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
-            WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
-            QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-            IS_MASK=True,
-            MASK_CAUSAL=False,
-            MASK_LOCAL=False,
-            IS_LOGSIGMOID_GATE=IS_LOGSIGMOID_GATE,
-            CHECK_INF=True,
         )
 
         n_block_max_no_mask = n_block_max - 1
@@ -719,88 +513,35 @@ def _fwd_gated_base_kernel(
             block_shape=(TILE_N, TILE_K),
             order=(1, 0),
         )
-        d_ptrs = tl.make_block_ptr(
-            base=d_base,
-            shape=(actual_seqlen_k,),
-            strides=(1,),
-            offsets=((n_block_max_no_mask - 1) * TILE_N,),
-            block_shape=(TILE_N,),
-            order=(0,),
-        )
-
-        # Load key tile
         k_tile = tl.load(k_ptrs, boundary_check=(0, 1))
-
-        # Load delta tile
-        d_tile = tl.load(d_ptrs, boundary_check=(0,)).to(tl.float32)
-        d_max = tl.max(d_tile)
-        d_min = tl.min(d_tile)
-
-        # Check if any gates are active for current tile
-        gate_max, skip_gate_curr = activations.online_gate(
-            a_max,
-            a_min,
-            d_max,
-            d_min,
-            gate_max,
-            scale_log2=softmax_scale_log2,
-            gate_threshold_log2=gate_threshold_log2,
-        )
-
-        # Compute attention gates
-        acc_s = a_tile[:, None] * d_tile[None, :]
-
-        if IS_LOGSIGMOID_GATE:
-            acc_s = activations.log_sigmoid(acc_s, FASTMATH=True)
-
-        # Compute attention scores
-        acc_s += tl.dot(q_tile, k_tile)
-
         for n_block in tl.range(n_block_max_no_mask - 1, n_block_min_no_mask - 1, -1):
-            (
-                skip_gate_curr,
-                acc_s,
-                acc_o,
-                k_ptrs,
-                v_ptrs,
-                d_ptrs,
-                gate_max,
-                block_max,
-                row_max,
-                row_sum,
-            ) = _fwd_inner_gated_base_kernel(
-                skip_gate_curr=skip_gate_curr,
-                acc_s=acc_s,
-                acc_o=acc_o,
-                q_tile=q_tile,
-                a_tile=a_tile,
-                k_ptrs=k_ptrs,
-                v_ptrs=v_ptrs,
-                d_ptrs=d_ptrs,
-                a_max=a_max,
-                a_min=a_min,
-                gate_max=gate_max,
-                block_max=block_max,
-                row_max=row_max,
-                row_sum=row_sum,
-                softmax_scale_log2=softmax_scale_log2,
-                gate_threshold_log2=gate_threshold_log2,
-                softmax_threshold_log2=softmax_threshold_log2,
-                m_block=m_block,
-                n_block=n_block,
-                n_block_min=n_block_min_no_mask,
-                actual_seqlen_q=actual_seqlen_q,
-                actual_seqlen_k=actual_seqlen_k,
-                TILE_M=TILE_M,
-                TILE_N=TILE_N,
-                WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
-                WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
-                QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-                IS_MASK=IS_LOCAL,
-                MASK_CAUSAL=False,
-                MASK_LOCAL=False,
-                IS_LOGSIGMOID_GATE=IS_LOGSIGMOID_GATE,
-                CHECK_INF=True,
+            k_tile, k_ptrs, v_ptrs, acc_o, block_max, row_max, row_sum = (
+                _fwd_inner_sparse_base_kernel(
+                    q_tile=q_tile,
+                    k_tile=k_tile,
+                    k_ptrs=k_ptrs,
+                    v_ptrs=v_ptrs,
+                    acc_o=acc_o,
+                    block_max=block_max,
+                    row_max=row_max,
+                    row_sum=row_sum,
+                    softmax_scale_log2=softmax_scale_log2,
+                    softmax_threshold_log2=softmax_threshold_log2,
+                    m_block=m_block,
+                    n_block=n_block,
+                    n_block_min=n_block_min_no_mask,
+                    actual_seqlen_q=actual_seqlen_q,
+                    actual_seqlen_k=actual_seqlen_k,
+                    TILE_M=TILE_M,
+                    TILE_N=TILE_N,
+                    WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
+                    WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
+                    QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
+                    IS_MASK=IS_LOCAL,
+                    MASK_CAUSAL=False,
+                    MASK_LOCAL=False,
+                    CHECK_INF=IS_LOCAL,
+                )
             )
 
     # Process n_blocks with masking
@@ -821,88 +562,35 @@ def _fwd_gated_base_kernel(
             block_shape=(TILE_N, TILE_K),
             order=(1, 0),
         )
-        d_ptrs = tl.make_block_ptr(
-            base=d_base,
-            shape=(actual_seqlen_k,),
-            strides=(1,),
-            offsets=((n_block_min_no_mask - 1) * TILE_N,),
-            block_shape=(TILE_N,),
-            order=(0,),
-        )
-
-        # Load key tile
         k_tile = tl.load(k_ptrs, boundary_check=(0, 1))
-
-        # Load delta tile
-        d_tile = tl.load(d_ptrs, boundary_check=(0,)).to(tl.float32)
-        d_max = tl.max(d_tile)
-        d_min = tl.min(d_tile)
-
-        # Check if any gates are active for current tile
-        gate_max, skip_gate_curr = activations.online_gate(
-            a_max,
-            a_min,
-            d_max,
-            d_min,
-            gate_max,
-            scale_log2=softmax_scale_log2,
-            gate_threshold_log2=gate_threshold_log2,
-        )
-
-        # Compute attention gates
-        acc_s = a_tile[:, None] * d_tile[None, :]
-
-        if IS_LOGSIGMOID_GATE:
-            acc_s = activations.log_sigmoid(acc_s, FASTMATH=True)
-
-        # Compute attention scores
-        acc_s += tl.dot(q_tile, k_tile)
-
         for n_block in tl.range(n_block_min_no_mask - 1, n_block_min - 1, -1):
-            (
-                skip_gate_curr,
-                acc_s,
-                acc_o,
-                k_ptrs,
-                v_ptrs,
-                d_ptrs,
-                gate_max,
-                block_max,
-                row_max,
-                row_sum,
-            ) = _fwd_inner_gated_base_kernel(
-                skip_gate_curr=skip_gate_curr,
-                acc_s=acc_s,
-                acc_o=acc_o,
-                q_tile=q_tile,
-                a_tile=a_tile,
-                k_ptrs=k_ptrs,
-                v_ptrs=v_ptrs,
-                d_ptrs=d_ptrs,
-                a_max=a_max,
-                a_min=a_min,
-                gate_max=gate_max,
-                block_max=block_max,
-                row_max=row_max,
-                row_sum=row_sum,
-                softmax_scale_log2=softmax_scale_log2,
-                gate_threshold_log2=gate_threshold_log2,
-                softmax_threshold_log2=softmax_threshold_log2,
-                m_block=m_block,
-                n_block=n_block,
-                n_block_min=n_block_min,
-                actual_seqlen_q=actual_seqlen_q,
-                actual_seqlen_k=actual_seqlen_k,
-                TILE_M=TILE_M,
-                TILE_N=TILE_N,
-                WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
-                WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
-                QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
-                IS_MASK=True,
-                MASK_CAUSAL=False,
-                MASK_LOCAL=True,
-                IS_LOGSIGMOID_GATE=IS_LOGSIGMOID_GATE,
-                CHECK_INF=True,
+            k_tile, k_ptrs, v_ptrs, acc_o, block_max, row_max, row_sum = (
+                _fwd_inner_sparse_base_kernel(
+                    q_tile=q_tile,
+                    k_tile=k_tile,
+                    k_ptrs=k_ptrs,
+                    v_ptrs=v_ptrs,
+                    acc_o=acc_o,
+                    block_max=block_max,
+                    row_max=row_max,
+                    row_sum=row_sum,
+                    softmax_scale_log2=softmax_scale_log2,
+                    softmax_threshold_log2=softmax_threshold_log2,
+                    m_block=m_block,
+                    n_block=n_block,
+                    n_block_min=n_block_min,
+                    actual_seqlen_q=actual_seqlen_q,
+                    actual_seqlen_k=actual_seqlen_k,
+                    TILE_M=TILE_M,
+                    TILE_N=TILE_N,
+                    WINDOW_SIZE_LEFT=WINDOW_SIZE_LEFT,
+                    WINDOW_SIZE_RIGHT=WINDOW_SIZE_RIGHT,
+                    QHEADS_PER_KVHEAD_PACKGQA=QHEADS_PER_KVHEAD_PACKGQA,
+                    IS_MASK=True,
+                    MASK_CAUSAL=False,
+                    MASK_LOCAL=True,
+                    CHECK_INF=True,
+                )
             )
 
     # Finalize softmax
@@ -941,21 +629,16 @@ def _fwd_gated_base_kernel(
         tl.store(out_ptrs, acc_o, boundary_check=(0, 1))
 
 
-def _flash_gated_attn_base_forward(
+def _flash_sparse_attn_base_forward(
     query: torch.Tensor,
     key: torch.Tensor,
     value: torch.Tensor,
-    alpha: torch.Tensor,
-    delta: torch.Tensor,
     is_causal: bool = False,
     softmax_scale: float = None,
     softmax_threshold: float = None,
-    gate_threshold: float = None,
-    is_logsigmoid_gate: bool = False,
-    is_adapt_gate: bool = True,
     window_size: Tuple[int, int] = (None, None),
     pack_gqa: bool = False,
-) -> Tuple[torch.Tensor, torch.Tensor, float, float, float]:
+) -> Tuple[torch.Tensor, torch.Tensor, float, float]:
     num_SMs = torch.cuda.get_device_properties(query.device).multi_processor_count
     batch_size, seqlen_q, num_heads_q, head_dim = query.shape
     _, seqlen_k, num_heads_kv, _ = key.shape
@@ -965,7 +648,6 @@ def _flash_gated_attn_base_forward(
     softmax_scale = softmax_scale or 1.0 / (head_dim**0.5)
     softmax_scale_log2 = softmax_scale * math.log2(math.e)
     softmax_threshold = softmax_threshold or head_dim / seqlen_k
-    gate_threshold = gate_threshold or head_dim / seqlen_k
     qheads_per_kvhead = num_heads_q // num_heads_kv
     qheads_per_kvhead_packgqa = num_heads_q // num_heads_kv if pack_gqa else 1
 
@@ -973,8 +655,6 @@ def _flash_gated_attn_base_forward(
         query,
         key,
         value,
-        alpha=alpha,
-        delta=delta,
         cu_seqlens_q=None,
         cu_seqlens_k=None,
         num_heads_q=num_heads_q,
@@ -1005,11 +685,11 @@ def _flash_gated_attn_base_forward(
         else 1
     )
 
-    out = torch.zeros_like(query)
-    lse = torch.zeros(
+    out = torch.empty_like(query)
+    lse = torch.empty(
         (batch_size, num_heads_q, seqlen_q),
-        device=query.device,
         dtype=torch.float32,
+        device=query.device,
     )
 
     if is_split_kv:
@@ -1033,17 +713,14 @@ def _flash_gated_attn_base_forward(
         num_splits=num_splits,
     )
 
-    _fwd_gated_base_kernel[grid](
+    _fwd_base_sparse_kernel[grid](
         query,
         key,
         value,
-        alpha,
-        delta,
         out if not is_split_kv else out_partial,
         lse if not is_split_kv else lse_partial,
         softmax_scale_log2,
         softmax_threshold,
-        gate_threshold,
         query.stride(0),
         query.stride(-2),
         query.stride(-3),
@@ -1053,10 +730,6 @@ def _flash_gated_attn_base_forward(
         value.stride(0),
         value.stride(-2),
         value.stride(-3),
-        alpha.stride(0),
-        alpha.stride(-2),
-        delta.stride(0),
-        delta.stride(-2),
         out.stride(0) if not is_split_kv else out_partial.stride(1),
         out.stride(-2) if not is_split_kv else out_partial.stride(-2),
         out.stride(-3) if not is_split_kv else out_partial.stride(-3),
@@ -1087,30 +760,26 @@ def _flash_gated_attn_base_forward(
         HAS_SEQUSED_Q=False,
         HAS_SEQUSED_K=False,
         PACK_GQA=pack_gqa,
-        IS_LOGSIGMOID_GATE=is_logsigmoid_gate,
-        IS_ADAPT_GATE=is_adapt_gate,
         num_warps=num_warps,
         num_stages=num_stages,
         num_ctas=num_ctas,
     )
 
     if is_split_kv:
-        flash_decode_combine._flash_attn_fwd_combine(
+        flash_fwd_combine._flash_attn_fwd_combine(
             out_partial,
             lse_partial,
             out,
             lse,
         )
 
-    return out, lse, softmax_scale, softmax_threshold, gate_threshold
+    return out, lse, softmax_scale, softmax_threshold
 
 
-def _flash_gated_attn_varlen_base_forward(
+def _flash_sparse_attn_varlen_base_forward(
     query: torch.Tensor,
     key: torch.Tensor,
     value: torch.Tensor,
-    alpha: torch.Tensor,
-    delta: torch.Tensor,
     cu_seqlens_q: torch.Tensor,
     cu_seqlens_k: torch.Tensor,
     max_seqlen_q: int,
@@ -1118,12 +787,9 @@ def _flash_gated_attn_varlen_base_forward(
     is_causal: bool = False,
     softmax_scale: float = None,
     softmax_threshold: float = None,
-    gate_threshold: float = None,
-    is_logsigmoid_gate: bool = True,
-    is_adapt_gate: bool = True,
     window_size: Tuple[int, int] = (None, None),
     pack_gqa: bool = False,
-) -> Tuple[torch.Tensor, torch.Tensor, float, float, float]:
+) -> Tuple[torch.Tensor, torch.Tensor, float, float]:
     num_SMs = torch.cuda.get_device_properties(query.device).multi_processor_count
     total_seqlen_q, num_heads_q, head_dim = query.shape
     _, num_heads_kv, _ = key.shape
@@ -1136,7 +802,6 @@ def _flash_gated_attn_varlen_base_forward(
     softmax_scale = softmax_scale or 1.0 / (head_dim**0.5)
     softmax_scale_log2 = softmax_scale * math.log2(math.e)
     softmax_threshold = softmax_threshold or head_dim / seqlen_k
-    gate_threshold = gate_threshold or head_dim / seqlen_k
     qheads_per_kvhead = num_heads_q // num_heads_kv
     qheads_per_kvhead_packgqa = num_heads_q // num_heads_kv if pack_gqa else 1
 
@@ -1144,8 +809,6 @@ def _flash_gated_attn_varlen_base_forward(
         query,
         key,
         value,
-        alpha=alpha,
-        delta=delta,
         cu_seqlens_q=cu_seqlens_q,
         cu_seqlens_k=cu_seqlens_k,
         num_heads_q=num_heads_q,
@@ -1176,7 +839,7 @@ def _flash_gated_attn_varlen_base_forward(
         else 1
     )
 
-    out = torch.zeros_like(query)
+    out = torch.empty_like(query)
     lse = torch.empty(
         (num_heads_q, total_seqlen_q),
         dtype=torch.float32,
@@ -1204,17 +867,14 @@ def _flash_gated_attn_varlen_base_forward(
         num_splits=num_splits,
     )
 
-    _fwd_gated_base_kernel[grid](
+    _fwd_base_sparse_kernel[grid](
         query,
         key,
         value,
-        alpha,
-        delta,
         out if not is_split_kv else out_partial,
         lse if not is_split_kv else lse_partial,
         softmax_scale_log2,
         softmax_threshold,
-        gate_threshold,
         0,
         query.stride(-2),
         query.stride(0),
@@ -1224,10 +884,6 @@ def _flash_gated_attn_varlen_base_forward(
         0,
         value.stride(-2),
         value.stride(0),
-        0,
-        alpha.stride(-2),
-        0,
-        delta.stride(-2),
         0,
         out.stride(-2) if not is_split_kv else out_partial.stride(-2),
         out.stride(0) if not is_split_kv else out_partial.stride(-3),
@@ -1258,15 +914,13 @@ def _flash_gated_attn_varlen_base_forward(
         HAS_SEQUSED_Q=False,
         HAS_SEQUSED_K=False,
         PACK_GQA=pack_gqa,
-        IS_LOGSIGMOID_GATE=is_logsigmoid_gate,
-        IS_ADAPT_GATE=is_adapt_gate,
         num_warps=num_warps,
         num_stages=num_stages,
         num_ctas=num_ctas,
     )
 
     if is_split_kv:
-        flash_decode_combine._flash_attn_fwd_combine(
+        flash_fwd_combine._flash_attn_fwd_combine(
             out_partial,
             lse_partial,
             out,
@@ -1274,4 +928,4 @@ def _flash_gated_attn_varlen_base_forward(
             cu_seqlens_q=cu_seqlens_q,
         )
 
-    return out, lse, softmax_scale, softmax_threshold, gate_threshold
+    return out, lse, softmax_scale, softmax_threshold
